refactor: type 별도 파일로 분리
src/features/navigation/ui/NavSidebarIcon.tsx
src/features/userStatus/ui/UserAvatar.tsx
