+++ src/features/navigation/index.ts
@@ -1,3 +1,3 @@
 export { NavGroupList } from './ui/NavGroupList'
 export { NAV_ITEMS } from './model/navItems'
-export type { NavItem, NavGroup, UserPlan } from './model/types'
+export type { NavItem, NavGroup, UserPlan, IconName, SidebarIconProps } from './model/types'
