+++ src/features/userStatus/index.ts
@@ -2,3 +2,4 @@ export { UserStatusCard } from './ui/UserStatusCard'
 export { UserAvatar } from './ui/UserAvatar'
 export { PlanButton } from './ui/PlanButton'
 export { ChannelStatus } from './ui/ChannelStatus'
+export type { UserAvatarProps } from './model/types'
