+++ src/features/userStatus/model/types.ts
@@ -1,0 +1,4 @@
+export interface UserAvatarProps {
+  size?: 'sm' | 'default' | 'lg'
+  showBadge?: boolean
+}
